Merge request for customer

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -1,8 +1,6 @@
 package ru.yandex.practicum.filmorate.model;
 
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Value;
+import lombok.*;
 
 @Value
 @AllArgsConstructor
@@ -10,7 +8,7 @@ import lombok.Value;
 public class Genre extends BaseEntity {
     private String name;
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         super(id);
         this.name = name;
     }
